% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calibrate}
\alias{calibrate}
\title{Calibration across data-sets}
\usage{
calibrate(x)
}
\arguments{
\item{x}{An \code{MPG} object}
}
\value{
A list with three objects:
\code{Ycal} is a matrix containing the calibrated data;
\code{calibration_distribution} is an array
with the posterior distribution of the shift for each observation;
\code{calibration_median} is a matrix with the posterior median shift
for each observation.
}
\description{
This function calibrates data-sets by removing misalignment.
}
\examples{
n = c(250, 250)
p = 4

Y1 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 3, ncol = p))
Y2 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 4, ncol = p))
Y = rbind(Y1, Y2)
C = c( rep(1,sum(n)), rep(2,sum(n)))

ans = mpg(Y, C)
cal = calibrate(ans)
par(mfrow=c(1,2))
plot(Y, col = C)
plot(cal$Y_cal, col = C)
}

