\name{feCodebook}
\alias{feCodebook}
\title{Fuel economy codebook}
\description{View a list of variable names and descriptions for the vehicle fuel economy data}
\usage{
feCodebook(browse = FALSE)
}
\arguments{
  \item{browse}{A logical indicating whether to open the web-based codebok. Default is \code{FALSE}.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
