\name{feData}
\alias{feData}
\title{Bulk data}
\description{Download fuel economy bulk data files}
\usage{
feData(type, year = NULL, ...)
}
\arguments{
  \item{type}{The type of data file to retrieve.}
  \item{year}{Optionally, a two- or four-digit year after 1977.}
  \item{\dots}{Ignored.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
