\name{feMakes}
\alias{feMakes}
\title{Vehicle makes}
\description{Retrieve makes for a vehicle year}
\usage{
feMakes(year, ...)
}
\arguments{
  \item{year}{A character string containing a four-digit vehicle year.}
  \item{\dots}{Ignored.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
