\name{feModels}
\alias{feModels}
\title{Vehicle models}
\description{Retrieve models for a make and year}
\usage{
feModels(year, make, ...)
}
\arguments{
  \item{year}{A character string containing a four-digit vehicle year.}
  \item{make}{A character string containing a vehicle make.}
  \item{\dots}{Ignored.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
