\name{feOptions}
\alias{feOptions}
\title{Vehicle options}
\description{Retrieve vehicle options for a make-model-year}
\usage{
feOptions(year, make, model, ...)
}
\arguments{
  \item{year}{A character string containing a four-digit vehicle year.}
  \item{make}{A character string containing a vehicle make.}
  \item{model}{A character string containing a vehicle model.}
  \item{\dots}{Ignored.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
