\name{feQuery}
\alias{feQuery}
\title{API Query}
\description{Execute an API query}
\usage{
feQuery(op, baseurl = "http://www.fueleconomy.gov/ws/rest/", ...)
}
\arguments{
  \item{op}{A character string containing an API operation, and associated arguments.}
  \item{baseurl}{The base URL for conducting the API query.}
  \item{\dots}{Ignored.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
