\name{feVehicle}
\alias{feVehicle}
\title{Vehicle fuel economy}
\description{Retrieve vehicle fuel economy by vehicle ID}
\usage{
feVehicle(id = NULL, pick = FALSE, ...)
}
\arguments{
  \item{id}{A vehicle ID number.}
  \item{pick}{A logical indicating whether to select the ID number via an interactive menu.}
  \item{\dots}{Options passed to \code{\link{pickVehicle}}, if \code{pick=TRUE}.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
