% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mpg}
\alias{mpg}
\title{Fit Mixtures of Perturbed Gaussians}
\usage{
mpg(Y, C, prior = NULL, mcmc = NULL, state = NULL)
}
\arguments{
\item{Y}{Matrix of the data. Each row represents an observation.}

\item{C}{Vector of the group label of each observation. Labels are integers starting from 1.}

\item{prior}{A list giving the prior information. If unspecified, a default prior is used.
The list includes the following hyparameters:
\code{K_0} Number of mixture components with shared weights.
\code{K_1} Number of mixture components with independent weights.
\code{epsilon_range} Vector with minimum and maximum values for \code{epsilon}.
\code{merge_step} Introduce step to merge mixture components with small KL divergence. Default is \code{merge_step = TRUE}.
\code{merge_par} Parameter controlling merging radius. Default is \code{merge_par = 0.1}.}

\item{mcmc}{A list giving the MCMC parameters. If unspecified, default parameters are used.
The list includes the following parameters: \code{nburn} indicates the number of burn-in scans,
\code{nsave} indicates the number of scans to be saved,
\code{nskip} indicates the thinning interval,
\code{ndisplay} indicates the number of scans to be displayed on screen.
The total number of scans is equal to  \code{nburn + nsave*nskip}.}

\item{state}{Initial state of the chain. At the moment only the latent variables Z can be initialized.}
}
\value{
A \code{MPG} object.
}
\description{
This function generates a sample from the posterior of Mixtures of Perturbed Gaussians.
}
\details{
\deqn{y_{i,j} = \sum_{k=1}^{K_0+K_1}\pi_{j,k}N(y_{i,j} | \mu_{j,k}, \Sigma_k )  }
where \eqn{i = 1, \ldots, n_j} and \eqn{j = 1, \ldots, J}.
The weights are defined as follows:
\deqn{ \pi_{j,k} = \rho w_{0,k}   \;\;\;  k = 1, \ldots, K_0,}
\deqn{ \pi_{j,k} = (1 - \rho) w_{j,k + K_0}   \;\;\;  k = 1, \ldots, K_1, }
where
\deqn{(w_{0,1}, \ldots, w_{0,K_0}) \sim Dirichlet(\alpha_0/K_0) }
\deqn{(w_{j,1}, \ldots, w_{j,K_1}) \sim Dirichlet(\alpha_j/K_1) }
}
\examples{
n = c(250, 250)
p = 4

Y1 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 3, ncol = p))
Y2 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 4, ncol = p))
Y = rbind(Y1, Y2)
C = c( rep(1,sum(n)), rep(2,sum(n)))

ans = mpg(Y, C)
}

