\name{mympgAvg}
\alias{mympgAvg}
\title{My MPG averages}
\description{Retrieve average fuel economy from My MPG by vehicle ID}
\usage{
mympgAvg(id = NULL, pick = FALSE, ...)
}
\arguments{
  \item{id}{A vehicle ID number.}
  \item{pick}{A logical indicating whether to select the ID number via an interactive menu.}
  \item{\dots}{Options passed to \code{\link{pickVehicle}}, if \code{pick=TRUE}.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
