\name{mympgModels}
\alias{mympgModels}
\title{My MPG models}
\description{Retrieve My MPG models for a given make}
\usage{
mympgModels(make, ...)
}
\arguments{
  \item{make}{A character string containing a vehicle make.}
  \item{\dots}{Ignored.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
