\name{mympgVehicles}
\alias{mympgVehicles}
\title{My MPG vehicles}
\description{Retrieve My MPG vehicle data by make and model}
\usage{
mympgVehicles(make, model, ...)
}
\arguments{
  \item{make}{A character string containing a vehicle make.}
  \item{model}{A character string containing a vehicle model.}
  \item{\dots}{Ignored.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
