\name{pickMyMPG}
\alias{pickMyMPG}
\title{Interactive vehicle menus}
\description{Use interactive menus to retrieve My MPG data}
\usage{
pickMyMPG(graphics = TRUE)
}
\arguments{
  \item{graphics}{A logical indicating whether to use graphical menus (the default) or, if \code{FALSE}, text-based menus.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
