% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDiff}
\alias{plotDiff}
\title{Plot differences across mixture models}
\usage{
plotDiff(object, type = "weight", dim = c(1, 2),
  group = 1:length(unique(object$data$C)), xlim = range(object$data$Y[,
  dim[1]]), ylim = range(object$data$Y[, dim[2]]), ...)
}
\arguments{
\item{object}{Object of class inheriting from "mpg".}
}
\value{
A plot.
}
\description{
Plot differences across mixture models
}
\examples{
n = c(250, 250)
p = 4

Y1 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 3, ncol = p))
Y2 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 4, ncol = p))
Y = rbind(Y1, Y2)
C = c( rep(1,sum(n)), rep(2,sum(n)))

ans = mpg(Y, C)
plotDiff(ans, type = "weight")
plotDiff(ans, type = "shift")
}

