\name{mpg-package}
\alias{mpg-package}
\alias{mpg}
\docType{package}
\title{FuelEconomy.gov Data}
\description{Extract fuel economy data from FuelEconomy.gov}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{\url{http://www.fueleconomy.gov/feg/ws/index.shtml}}
\keyword{ package }
%\seealso{}
%\examples{}
