\name{mympgVehicles}
\alias{mympgVehicles}
\title{My MPG vehicles}
\description{Retrieve My MPG vehicle data by make and model}
\usage{
mympgVehicles(make, model, pick=FALSE, graphics=TRUE, ...)
}
\arguments{
  \item{make}{A character string containing a vehicle make.}
  \item{model}{A character string containing a vehicle model.}
  \item{pick}{A logical indicating whether to select the ID number via an interactive menu.}
  \item{graphics}{If \code{pick=TRUE}, a logical indicating whether to use graphical menus (the default) or, if \code{FALSE}, text-based menus.}
  \item{\dots}{Ignored.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
